Add files via upload
This file contains the code for a very simple and basic streamlit chatbot. 
The LLM used is Llama3-70b-8192. It currently has access to data of 5 case reports in txt format. Limitations: No system prompt was given to the model; the chatbot does not have memory. Will work on these in the next iteration.